Update readme, preset and fix some build issues

+++ lib/gl/renderer/softwarerasterizer/Framebuffer.hpp
@@ -72,10 +72,10 @@ public:
         const int32_t tileWidth = static_cast<int32_t>(m_resolutionData.x);
         const int32_t yOff = static_cast<int32_t>(m_yOffset);
         // Scissor bounds are in screen space — convert Y to local tile space by subtracting yOffset
-        const std::size_t startX = (m_scissorData.enabled) ? static_cast<std::size_t>(std::clamp(m_scissorData.startX, 0, tileWidth)) : 0;
-        const std::size_t endX = (m_scissorData.enabled) ? static_cast<std::size_t>(std::clamp(m_scissorData.endX, 0, tileWidth)) : m_resolutionData.x;
-        const std::size_t startY = (m_scissorData.enabled) ? static_cast<std::size_t>(std::clamp(m_scissorData.startY - yOff, 0, tileHeight)) : 0;
-        const std::size_t endY = (m_scissorData.enabled) ? static_cast<std::size_t>(std::clamp(m_scissorData.endY - yOff, 0, tileHeight)) : m_resolutionData.y;
+        const std::size_t startX = (m_scissorData.enabled) ? static_cast<std::size_t>(std::clamp(m_scissorData.startX, static_cast<int32_t>(0), tileWidth)) : static_cast<std::size_t>(0);
+        const std::size_t endX = (m_scissorData.enabled) ? static_cast<std::size_t>(std::clamp(m_scissorData.endX, static_cast<int32_t>(0), tileWidth)) : m_resolutionData.x;
+        const std::size_t startY = (m_scissorData.enabled) ? static_cast<std::size_t>(std::clamp(m_scissorData.startY - yOff, static_cast<int32_t>(0), tileHeight)) : static_cast<std::size_t>(0);
+        const std::size_t endY = (m_scissorData.enabled) ? static_cast<std::size_t>(std::clamp(m_scissorData.endY - yOff, static_cast<int32_t>(0), tileHeight)) : m_resolutionData.y;
         for (std::size_t y = startY; y < endY; y++)
         {
             for (std::size_t x = startX; x < endX; x++)
