+++ lib/gl/renderer/softwarerasterizer/TextureMap.hpp
@@ -129,8 +129,8 @@ private:
 
         if (m_wrapModeS == TextureWrapMode::CLAMP_TO_EDGE)
         {
-            return { static_cast<uint32_t>(std::clamp(sInt, 0, static_cast<int32_t>(m_textureMaskW))),
-                static_cast<uint32_t>(std::clamp(tInt, 0, static_cast<int32_t>(m_textureMaskH))) };
+            return { static_cast<uint32_t>(std::clamp(sInt, static_cast<int32_t>(0), static_cast<int32_t>(m_textureMaskW))),
+                static_cast<uint32_t>(std::clamp(tInt, static_cast<int32_t>(0), static_cast<int32_t>(m_textureMaskH))) };
         }
         // REPEAT - use bitmask for power-of-2 textures
         return { static_cast<uint32_t>(sInt) & m_textureMaskW,
